+++ management/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.36.0 (Unreleased)
 
+* **Breaking Change** Due to validation on the API, `EnumNotificationsPolicyQuotaDeliveryMethods` now becomes `NotificationsPolicyQuotasInnerDeliveryMethodsInner` [#312](https://github.com/patrickcping/pingone-go-sdk-v2/pull/312)
 * **Enhancement** Add ability to designate an application's client secret as a previous secret with an expiry date up to 30 days, for rotation purposes. [#311](https://github.com/patrickcping/pingone-go-sdk-v2/pull/311)
 
 # v0.35.0 (2024-01-12)